requirements.txt
asn1crypto==0.24.0
Authlib==0.11
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.2
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
google-auth==1.6.3
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
python-decouple==3.1
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
